Refactor code-examples and variables preview into markdown renderer rule
